Fix double-registration of mapper modules
The Resque :preload task wants to eagerly load our application's
classes:
https://github.com/resque/resque/blob/1-x-stable/lib/resque/tasks.rb#L62

For reasons that aren't yet clear, this was causing the
Krikri::Mapper.define call in esdn_mods.rb to happen twice.  This, of
course, was throwing an exception.  The application was loading
correctly in the Rails console, but not in the resque:work rake task.

This change introduces a simple conditional to call #define only if the
mapping is unregistered.