Upgrade Rake & tighten up rake dependency
Rake 11.0.0 was recently released. This upgrades us explicitly and locks
down the dependency to the minor version for better future safety.

We're doing this on Heidrun rather than Krikri to avoid locking other
potential Krikri (engine) users to a specific Rake version. This is more
maintenance overhead on Krikri in the long run, but it seems like the
only viable choice to avoid dependency conflicts.

Updates the RSpec dependency to 3.4.4 for Rake 11 compatibility.